Move remaining TdsParse clases wholesale Move TdsParserStateObjectManaged to Parser namespace
