+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/Parser/TdsParserStaticMethods.cs
@@ -7,7 +7,7 @@ using System.Globalization;
 using Microsoft.Data.Common;
 using Microsoft.Data.SqlClient.Parser;
 
-namespace Microsoft.Data.SqlClient
+namespace Microsoft.Data.SqlClient.Parser
 {
     internal sealed class TdsParserStaticMethods
     {
